+++ include/steps/fastq_screen.py
@@ -13,7 +13,8 @@ class FastqScreen(AbstractStep):
 
     FastQ Screen allows you to screen a library of sequences in FastQ format
     against a set of sequence databases so you can see if the composition of
-    the library matches with what you expect.
+    the library matches with what you expect. This step uses Bowtie2 as the
+    alignment tool.
 
     '''