+++ CHANGELOG.md
@@ -86,6 +86,7 @@
  * git status --porcelain output with changes and untracked files in annotation files
  * more info on running task in `status <task name>`
  * include the script `collect_scs.py` in the uap (#164)
+ * `_pattern` and `_replacement` step options to rename run ids
 
 **additional stuff**
  * updated documentation and resolved sphinx warnings